+++ linwalker/__init__.py
@@ -1,4 +1,4 @@
 """LINwalker: explore hierarchical population structure and introgression using LIN codes."""
 
 __all__ = ["__version__"]
-__version__ = "1.0.17"
+__version__ = "1.0.18"